Bug fixes - Buttons does not lag after update and info message does not show error state after update.
