+++ plugins/ros/src/components/scenarioDrawer/components/ActionsSection.tsx
@@ -4,7 +4,7 @@ import Paper from '@mui/material/Paper';
 import Typography from '@mui/material/Typography';
 import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
 import { pluginRiScTranslationRef } from '../../../utils/translations';
-import { emptyAction, useScenario } from '../../../contexts/ScenarioContext';
+import { emptyAction } from '../../../contexts/ScenarioContext';
 import { section } from '../scenarioDrawerComponents';
 import { emptyState, heading3 } from '../../common/typography';
 import Divider from '@mui/material/Divider';
@@ -16,7 +16,6 @@ import { AddCircle } from '@mui/icons-material';
 import Box from '@mui/material/Box';
 import { ActionStatusOptions } from '../../../utils/constants';
 import Switch from '@mui/material/Switch';
-import { useDebounce } from '../../../utils/hooks';
 
 const FILTER_SETTINGS = {
   SHOW_ALL: false,
@@ -27,6 +26,7 @@ type ActionSectionProps = {
   formMethods: UseFormReturn<FormScenario>;
   isEditing: boolean;
   onSubmit: () => void;
+  setCurrentUpdatedActionIDs: React.Dispatch<React.SetStateAction<string[]>>;
 };
 
 const RelevanceToggle = ({
@@ -57,6 +57,7 @@ export function ActionsSection({
   formMethods,
   isEditing,
   onSubmit,
+  setCurrentUpdatedActionIDs,
 }: ActionSectionProps) {
   const { t } = useTranslationRef(pluginRiScTranslationRef);
 
@@ -109,46 +110,6 @@ export function ActionsSection({
       originalIndex: currentActions.findIndex(a => a === action),
     }));
   }, [currentActions, showOnlyRelevant, filterActions, sortActionsByRelevance]);
-
-  const [currentUpdatedActionIDs, setCurrentUpdatedActionIDs] = useState<
-    string[]
-  >([]);
-
-  const { submitEditedScenarioToRiSc, scenario } = useScenario();
-
-  const debounceCallback = useCallback(
-    (updatedIDs: string[]) => {
-      const indexOfAction = (ID: string) => {
-        return scenario.actions.findIndex(a => a.ID === ID);
-      };
-      if (updatedIDs.length === 0) return;
-
-      const formValues = formMethods.getValues();
-      const updatedScenario = {
-        ...scenario,
-        actions: scenario.actions.map(a =>
-          updatedIDs.includes(a.ID)
-            ? {
-                ...a,
-                status:
-                  formValues.actions?.[indexOfAction(a.ID)]?.status ?? a.status,
-                lastUpdated: new Date(),
-              }
-            : a,
-        ),
-      };
-      submitEditedScenarioToRiSc(updatedScenario);
-      setCurrentUpdatedActionIDs([]);
-    },
-    [
-      scenario,
-      formMethods,
-      setCurrentUpdatedActionIDs,
-      submitEditedScenarioToRiSc,
-    ],
-  );
-
-  useDebounce(currentUpdatedActionIDs, 6000, debounceCallback);
 
   if (isEditing) {
     return (
